server/pyproject.toml
[tool.poetry]
name = "blog-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.16.0"
django-environ = "^0.12.0"
ruff = "^0.11.2"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
